navigasi fix
