+++ lib/main.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:rojotani/Awal/loginAs.dart';
-import 'package:rojotani/Awal/loginPenjual.dart';
+import 'package:rojotani/Awal/loginPetani.dart';
 import 'package:rojotani/Awal/registerAs.dart';
-import 'package:rojotani/pembeli/transaksi/cekOut.dart';
-import 'package:rojotani/pembeli/transaksi/pengiriman.dart';
+import 'package:rojotani/pelanggan/transaksi/cekOut.dart';
+import 'package:rojotani/pelanggan/transaksi/pengiriman.dart';
 import 'package:rojotani/petani/akun/akunPetani.dart';
 import 'package:rojotani/petani/akun/katasandi.dart';
 import 'package:rojotani/petani/navPetani.dart';
@@ -12,7 +12,7 @@ import 'package:rojotani/petani/produk/tambah_produk/tambahLelang.dart';
 import 'package:rojotani/petani/produk/tambah_produk/tambahProduk.dart';
 import 'package:rojotani/petani/produk/tambah_produk/tambahProduk.dart';
 import 'Awal/registerPelanggan.dart';
-import 'pembeli/produk/home.dart';
+import 'pelanggan/produk/home.dart';
 import 'petani/transaksi/keranjang.dart';
 import 'SplashScreen.dart';
 import 'Awal/loginPelanggan.dart';
@@ -32,7 +32,7 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         scaffoldBackgroundColor: Colors.white,
       ),
-      home: loginPelangganPage(),
+      home: loginAsPage(),
       //homePage()
     );
   }
