+++ lib/pelanggan/produk/akun/akunPetani.dart
@@ -3,19 +3,20 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:image_picker/image_picker.dart';
+import 'package:rojotani/Awal/loginPelanggan.dart';
 import 'package:rojotani/Awal/loginPetani.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-class akunPetani extends StatefulWidget {
-  // const akunPetani({Key key}) : super(key: key);
+class akunPelanggan extends StatefulWidget {
+  // const akunPelanggan({Key key}) : super(key: key);
 
   @override
   // final VoidCallback signOut;
-  // akunPetani(this.signOut);
-  State<akunPetani> createState() => _akunPetaniState();
+  // akunPelanggan(this.signOut);
+  State<akunPelanggan> createState() => _akunPelangganState();
 }
 
-class _akunPetaniState extends State<akunPetani> {
+class _akunPelangganState extends State<akunPelanggan> {
   File _image;
 
   Future getImage() async {
@@ -29,7 +30,7 @@ class _akunPetaniState extends State<akunPetani> {
   signOut() async {
     SharedPreferences preferences = await SharedPreferences.getInstance();
     preferences.clear();
-    Route route = MaterialPageRoute(builder: (context) => loginPetaniPage());
+    Route route = MaterialPageRoute(builder: (context) => loginPelangganPage());
     Navigator.push(context, route);
   }
